+++ example/lib/research_package_objects/informed_consent.dart
@@ -183,6 +183,8 @@ RPOrderedTask get consentTask {
     title: 'Informed Consent - Full Version',
     sections: [
       overviewSection,
+      welcomeSection,
+      aboutUs,
       dataGatheringSection,
       privacySection,
       dataUseSection,
@@ -192,8 +194,6 @@ RPOrderedTask get consentTask {
       studySurveySection,
       withdrawingSection,
       yourRightsSection,
-      welcomeSection,
-      aboutUs,
       goalsSection,
       benefitsSection,
       locationSection,
@@ -222,6 +222,8 @@ RPOrderedTask get consentTask {
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam ultricies feugiat turpis nec efficitur. Integer in pharetra libero. Proin a leo eu enim porttitor hendrerit. Suspendisse vestibulum interdum mollis. Donec in sapien ut orci ultricies laoreet. Ut maximus ante id arcu feugiat scelerisque. Proin non rutrum libero. Aliquam blandit arcu ac dolor consequat maximus. Integer et dolor quis quam tempor porta quis vel nibh. Phasellus ullamcorper fringilla lorem, ac tempus sem cursus a. Aliquam maximus facilisis quam. Morbi hendrerit tempor tellus, ac hendrerit augue tincidunt eu. Cras convallis lorem at nulla mattis tristique.",
     footnote: "(1) Important footnote",
     imagePath: "assets/images/waving-hand.png",
+    // videoPath: "assets/video/colors.mp4",
+    audioPath: "audio/test_audio.mp3",
     text: "informed_consent.welcome_text",
   );