+++ lib/src/ui/completion_step.dart
@@ -34,7 +34,8 @@ class RPUICompletionStepState extends State<RPUICompletionStep>
   Widget build(BuildContext context) {
     RPLocalizations? locale = RPLocalizations.of(context);
     return Scaffold(
-      backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+      backgroundColor:
+          Theme.of(context).extension<CarpColors>()!.backgroundGray,
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -68,6 +69,7 @@ class RPUICompletionStepState extends State<RPUICompletionStep>
               ),
             ),
             ButtonTheme(
+              buttonColor: Theme.of(context).extension<CarpColors>()!.primary,
               minWidth: 150,
               child: OutlinedButton(
                 onPressed: () {