+++ lib/src/ui/instruction_step.dart
@@ -15,10 +15,35 @@ class RPUIInstructionStep extends StatefulWidget {
 }
 
 class RPUIInstructionStepState extends State<RPUIInstructionStep> {
+  late AudioPlayer player = AudioPlayer();
+
   @override
   void initState() {
     blocQuestion.sendReadyToProceed(true);
     super.initState();
+    player = AudioPlayer();
+    player.setReleaseMode(ReleaseMode.stop);
+
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      try {
+        if (widget.step.audioPath != null) {
+          if (widget.step.audioPath!.startsWith('http')) {
+            await player.setSource(UrlSource(widget.step.audioPath!));
+          } else {
+            await player.setSource(AssetSource(widget.step.audioPath!));
+          }
+        }
+      } catch (e) {
+        print(e);
+      }
+    });
+  }
+
+  @override
+  void dispose() {
+    // Release all sources and dispose the player.
+    player.dispose();
+    super.dispose();
   }
 
   void _pushDetailTextRoute() {
@@ -37,71 +62,81 @@ class RPUIInstructionStepState extends State<RPUIInstructionStep> {
   @override
   Widget build(BuildContext context) {
     RPLocalizations? locale = RPLocalizations.of(context);
-    return SafeArea(
-      child: Column(
-        children: <Widget>[
-          Expanded(
-            child: SingleChildScrollView(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  // If image is provided show it
-                  if (widget.step.imagePath != null)
-                    Center(
-                      child: SizedBox(
-                        height: MediaQuery.of(context).size.height * 0.25,
-                        width: MediaQuery.of(context).size.width * 0.25,
-                        child: InstructionImage(widget.step.imagePath!),
+    return Scaffold(
+      backgroundColor:
+          Theme.of(context).extension<CarpColors>()!.backgroundGray,
+      body: SafeArea(
+        child: Column(
+          children: <Widget>[
+            Expanded(
+              child: SingleChildScrollView(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    // If image is provided show it
+                    if (widget.step.imagePath != null)
+                      Center(
+                        child: SizedBox(
+                          height: MediaQuery.of(context).size.height,
+                          width: MediaQuery.of(context).size.width,
+                          child: InstructionImage(
+                              imagePath: widget.step.imagePath!),
+                        ),
                       ),
-                    ),
+                    if (widget.step.videoPath != null)
+                      VideoApp(step: widget.step),
+                    if (widget.step.audioPath != null)
+                      AudioPlayerWidget(player: player),
 
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: <Widget>[
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                            vertical: 30, horizontal: 16),
-                        child: Text(
-                          locale?.translate(widget.step.text!) ??
-                              widget.step.text!,
-                          textAlign: TextAlign.start,
-                          style: Theme.of(context).textTheme.headlineSmall,
-                        ),
-                      ),
-                      if (widget.step.detailText != null)
-                        TextButton(
-                          style: TextButton.styleFrom(
-                            textStyle: TextStyle(
-                              color: (CupertinoTheme.of(context).primaryColor ==
-                                      CupertinoColors.activeBlue)
-                                  ? Theme.of(context).primaryColor
-                                  : CupertinoTheme.of(context).primaryColor,
-                            ),
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: <Widget>[
+                        Container(
+                          padding: const EdgeInsets.symmetric(
+                              vertical: 30, horizontal: 16),
+                          child: Text(
+                            locale?.translate(widget.step.text!) ??
+                                widget.step.text!,
+                            textAlign: TextAlign.start,
+                            style: Theme.of(context).textTheme.headlineSmall,
                           ),
-                          onPressed: _pushDetailTextRoute,
-                          child: Text(locale?.translate('learn_more') ??
-                              "Learn more..."),
                         ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-          if (widget.step.footnote != null)
-            Center(
-              child: Container(
-                padding: const EdgeInsets.all(8.0),
-                child: Text(
-                  locale?.translate(widget.step.footnote!) ??
-                      widget.step.footnote!,
-                  style: Theme.of(context).textTheme.bodySmall,
-                  textAlign: TextAlign.center,
-                  softWrap: true,
+                        if (widget.step.detailText != null)
+                          TextButton(
+                            style: TextButton.styleFrom(
+                              textStyle: TextStyle(
+                                color: (CupertinoTheme.of(context)
+                                            .primaryColor ==
+                                        CupertinoColors.activeBlue)
+                                    ? Theme.of(context).primaryColor
+                                    : CupertinoTheme.of(context).primaryColor,
+                              ),
+                            ),
+                            onPressed: _pushDetailTextRoute,
+                            child: Text(locale?.translate('learn_more') ??
+                                "Learn more"),
+                          ),
+                      ],
+                    ),
+                  ],
                 ),
               ),
             ),
-        ],
+            if (widget.step.footnote != null)
+              Center(
+                child: Container(
+                  padding: const EdgeInsets.all(8.0),
+                  child: Text(
+                    locale?.translate(widget.step.footnote!) ??
+                        widget.step.footnote!,
+                    style: Theme.of(context).textTheme.bodySmall,
+                    textAlign: TextAlign.center,
+                    softWrap: true,
+                  ),
+                ),
+              ),
+          ],
+        ),
       ),
     );
   }
@@ -152,23 +187,6 @@ class _DetailTextRoute extends StatelessWidget {
             ),
           ],
         ),
-      ),
-    );
-  }
-}
-
-class InstructionImage extends StatelessWidget {
-  final String _imagePath;
-
-  const InstructionImage(this._imagePath, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(20.0),
-      child: Image.asset(
-        _imagePath,
-        width: MediaQuery.of(context).size.width / 2,
       ),
     );
   }