+++ lib/ui.dart
@@ -15,8 +15,10 @@ import 'dart:convert';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:just_audio/just_audio.dart';
 import 'package:signature/signature.dart';
+import 'package:audioplayers/audioplayers.dart';
+import 'package:video_player/video_player.dart';
+import 'package:carp_themes_package/carp_themes_package.dart';
 
 import 'model.dart';
 
@@ -35,6 +37,9 @@ part 'src/ui/completion_step.dart';
 part 'src/ui/consent_review_step.dart';
 part 'src/ui/form_step.dart';
 part 'src/ui/instruction_step.dart';
+part 'src/ui/instruction_step.audio_instruction.dart';
+part 'src/ui/instruction_step.image_instruction.dart';
+part 'src/ui/instruction_step.video_instruction.dart';
 part 'src/ui/question_step.dart';
 part 'src/ui/task.dart';
 part 'src/ui/visual_consent_step.dart';